requirements.txt
click==6.7
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
libsass==0.14.5
MarkupSafe==1.0
Pygments==2.2.0
redis==2.10.6
six==1.11.0
Werkzeug==0.14.1
